requirements.txt
flask
flask-sqlalchemy
flask-login
python-dotenv
mysql-connector-python
bcrypt
flask-wtf